Add files via upload
This is the code file containing the mathematical formulations of the game setting in Island Warrior and the optimization programming using Gurobi. 
The player of Island Warrior is expected to travel through a series of islands to collect weapons and eventually challenge the final boss. While the player can freely explore the islands, they are subject to various constraints of budgets, dexterity of using certain weapons, etc. 
The optimization is to find the quickiest route for the player to find the boss and finish the game with least amount of time.